[task-manager][expo-go] remove warn about expo go in task manager
Task Manager doesn't need to warn about being run on Expo Go, since the consumers of the service will warn about this themselves.

F.ex both Background Fetch and Background Tasks are available in Expo Go - while Location is not - but all of them are using the Task Manager.

The packages that isn't supported will warn individually.

This commit removes the tests for running on Expo Go in Task Manager. This is also mentioned here: #33776.